+++ src/components/Home/FeedList.tsx
@@ -5,13 +5,12 @@ import { HomeStackParams } from '../../pages/Home';
 import styled from 'styled-components/native';
 import { BLACK, LIGHTBLACK, MINT, WHITE } from '../../styles/GlobalColor';
 import { B14, B16, R14 } from '../../styles/GlobalText';
-import { BottomSheetModal, BottomSheetModalProvider, BottomSheetBackdrop } from "@gorhom/bottom-sheet";
+import {  BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 import { AppItem, UserItem } from './FeedItem';
-import BottomButton from '../common/BottomButton';
+import { CategoryModal, OrderModal } from '../common/BottomModal';
 
 import SearchIcon from '../../assets/common/Search.svg';
 import ArrowIcon from '../../assets/common/Arrow.svg';
-import CheckIcon from '../../assets/common/Check.svg';
 
 const UserData = [
   {
@@ -57,47 +56,10 @@ const FeedList = ({ navigation }: StackScreenProps<HomeStackParams, 'FeedList'>)
     { title: '00s', label: '00년대' },
     { title: '10s', label: '10년대' }
   ];
-
-  // 카테고리 모달
   const [categories, setCategories] = useState<string[]>([]);
-  const [prevCategories, setPrevCategories] = useState<string[]>([]);
-  const categoryRef = useRef<BottomSheetModal>(null);
-  const categorySnapPoints = useMemo(() => [350], []);
-
-  const openCategory = () => {
-    setPrevCategories([...categories]);
-    categoryRef.current?.present();
-  };
-
-  const closeCategory = () => {
-    if (categories.length === 0) {
-      setCategories([...prevCategories]);
-    }
-    categoryRef.current?.close();
-  }
-
-  const renderCategoryBackdrop = useCallback(
-    (props: any) => <BottomSheetBackdrop style={{flex: 1}} {...props} onPress={closeCategory} pressBehavior='close' appearsOnIndex={0} disappearsOnIndex={-1} />,
-    [],
-  );
-  
-  // 정렬 모달
+  const [categoryVisible, setCategoryVisible] = useState<boolean>(false);
   const [order, setOrder] = useState<string>('최신순');
-  const orderRef = useRef<BottomSheetModal>(null);
-  const orderSnapPoints = useMemo(() => [200], []);
-
-  const openOrder = () => {
-    orderRef.current?.present();
-  }
-
-  const closeOrder = () => {
-    orderRef.current?.close();
-  }
-
-  const renderOrderBackdrop = useCallback(
-    (props: any) => <BottomSheetBackdrop style={{flex: 1}} {...props} onPress={closeOrder} pressBehavior='close' appearsOnIndex={0} disappearsOnIndex={-1} />,
-    [],
-  );
+  const [orderVisible, setOrderVisible] = useState<boolean>(false);
 
   return (
     <SafeAreaView style={{flex: 1}}>
@@ -124,7 +86,7 @@ const FeedList = ({ navigation }: StackScreenProps<HomeStackParams, 'FeedList'>)
         ListHeaderComponent={() => {
           return (
             <View style={{flexDirection: 'row', paddingTop: 15, paddingBottom: 5, paddingHorizontal: 10}}>
-              <DropDownButton onPress={openCategory}>
+              <DropDownButton onPress={() => setCategoryVisible(true)}>
                 <R14 style={{color: BLACK, marginRight: 5}}>
                   {categories.length > 0 ? 
                     (categories.length > 1 ? `카테고리 ${categories.length}` : categories) 
@@ -134,7 +96,7 @@ const FeedList = ({ navigation }: StackScreenProps<HomeStackParams, 'FeedList'>)
                 </R14>
                 <ArrowIcon color={BLACK} />
               </DropDownButton>
-              <DropDownButton onPress={openOrder}>
+              <DropDownButton onPress={() => setOrderVisible(true)}>
                 <R14 style={{color: BLACK, marginRight: 5}}>{order}</R14>
                 <ArrowIcon color={BLACK} />
               </DropDownButton>
@@ -175,63 +137,8 @@ const FeedList = ({ navigation }: StackScreenProps<HomeStackParams, 'FeedList'>)
         }}
       />
       <BottomSheetModalProvider>
-        <BottomSheetModal
-          snapPoints={categorySnapPoints}
-          ref={categoryRef}
-          backdropComponent={renderCategoryBackdrop}
-          handleStyle={{backgroundColor: LIGHTBLACK, borderTopLeftRadius: 25, borderTopRightRadius: 25}}
-          handleIndicatorStyle={{backgroundColor: '#3F3F3F', width: 60}}
-          backgroundStyle={{backgroundColor: LIGHTBLACK}}
-          onDismiss={closeCategory}
-        >
-          <View style={{paddingHorizontal: 20, paddingVertical: 5}}>
-            <B16 style={{color: MINT, marginBottom: 10}}>카테고리</B16>
-            <FlatList
-              data={['영화', '드라마', '애니메이션', '패션', '음악', '예능']}
-              numColumns={2}
-              style={{marginBottom: 5}}
-              renderItem={({item}: any) => {
-                return (
-                  <CategoryButton pressed={categories.includes(item)}
-                    onPress={() => {
-                      categories.includes(item) ? 
-                      setCategories(categories.filter((category: string) => category !== item))
-                      : 
-                      setCategories([...categories, item])
-                    }}
-                  >
-                    <B16 style={{color: categories.includes(item) ? BLACK : WHITE}}>{item}</B16>
-                  </CategoryButton>
-                )
-              }}
-            />
-            <BottomButton label={'적용하기'} onPress={closeCategory} />
-          </View>
-        </BottomSheetModal>
-        <BottomSheetModal
-          snapPoints={orderSnapPoints}
-          ref={orderRef}
-          backdropComponent={renderOrderBackdrop}
-          handleStyle={{backgroundColor: LIGHTBLACK, borderTopLeftRadius: 25, borderTopRightRadius: 25}}
-          handleIndicatorStyle={{backgroundColor: '#3F3F3F', width: 60}}
-          backgroundStyle={{backgroundColor: LIGHTBLACK}}
-          onDismiss={closeOrder}
-        >
-          <View style={{paddingHorizontal: 20, paddingVertical: 5}}>
-            <B16 style={{color: MINT, marginBottom: 20}}>정렬</B16>
-            <FlatList
-              data={['최신순', '인기순']}
-              renderItem={({item}: any) => {
-                return (
-                  <OrderButton onPress={() => {setOrder(item); orderRef.current?.close();}}>
-                    <B16 style={{color: order===item ? MINT : WHITE}}>{item}</B16>
-                    {order===item && <CheckIcon />}
-                  </OrderButton>
-                )
-              }}
-            />
-          </View>
-        </BottomSheetModal>
+        <CategoryModal categories={categories} setCategories={setCategories} categoryVisible={categoryVisible} setCategoryVisible={setCategoryVisible} />
+        <OrderModal order={order} setOrder={setOrder} orderVisible={orderVisible} setOrderVisible={setOrderVisible} />
       </BottomSheetModalProvider>
     </SafeAreaView>
   )
